requirements.txt
flask
pandas
python-docx
pillow
requests